Update tests due to testing api changes
