+++ test/decoder/decoder_native_test.exs
@@ -4,8 +4,8 @@ defmodule Decoder.NativeTest do
   alias Membrane.Element.FFmpeg.H264.Decoder.Native, as: Dec
 
   test "Decode 1 240p frame" do
-    in_path = "fixtures/input-100-240p.h264" |> Path.expand(__DIR__)
-    ref_path = "fixtures/reference-100-240p.raw" |> Path.expand(__DIR__)
+    in_path = "../fixtures/input-100-240p.h264" |> Path.expand(__DIR__)
+    ref_path = "../fixtures/reference-100-240p.raw" |> Path.expand(__DIR__)
 
     assert {:ok, file} = File.read(in_path)
     assert {:ok, decoder_ref} = Dec.create()
