+++ test/decoder/decoding_test.exs
@@ -1,10 +1,12 @@
 defmodule DecoderTest do
-  alias Membrane.Pipeline
   use ExUnit.Case
+  import Membrane.Testing.Assertions
+  alias Membrane.Testing.Pipeline
+  alias Membrane.Element
 
   def prepare_paths(filename) do
-    in_path = "fixtures/input-#{filename}.h264" |> Path.expand(__DIR__)
-    reference_path = "fixtures/reference-#{filename}.raw" |> Path.expand(__DIR__)
+    in_path = "../fixtures/input-#{filename}.h264" |> Path.expand(__DIR__)
+    reference_path = "../fixtures/reference-#{filename}.raw" |> Path.expand(__DIR__)
     out_path = "/tmp/output-decoding-#{filename}.raw"
     File.rm(out_path)
     on_exit(fn -> File.rm(out_path) end)
@@ -12,7 +14,19 @@ defmodule DecoderTest do
   end
 
   def make_pipeline(in_path, out_path) do
-    Pipeline.start_link(DecodingPipeline, %{in: in_path, out: out_path, pid: self()}, [])
+    Pipeline.start_link(%Pipeline.Options{
+      elements: [
+        file_src: %Element.File.Source{chunk_size: 40_960, location: in_path},
+        parser: Element.FFmpeg.H264.Parser,
+        decoder: Element.FFmpeg.H264.Decoder,
+        sink: %Element.File.Sink{location: out_path}
+      ],
+      links: %{
+        {:file_src, :output} => {:parser, :input},
+        {:parser, :output} => {:decoder, :input},
+        {:decoder, :output} => {:sink, :input}
+      }
+    })
   end
 
   def assert_files_equal(file_a, file_b) do
@@ -28,7 +42,7 @@ defmodule DecoderTest do
       assert {:ok, pid} = make_pipeline(in_path, out_path)
 
       assert Pipeline.play(pid) == :ok
-      assert_receive :eos, 500
+      assert_end_of_stream(pid, :sink, :input, 500)
       assert_files_equal(out_path, ref_path)
     end
 
@@ -38,7 +52,7 @@ defmodule DecoderTest do
       assert {:ok, pid} = make_pipeline(in_path, out_path)
 
       assert Pipeline.play(pid) == :ok
-      assert_receive :eos, 1000
+      assert_end_of_stream(pid, :sink, :input, 1000)
       assert_files_equal(out_path, ref_path)
     end
 
@@ -48,7 +62,7 @@ defmodule DecoderTest do
       assert {:ok, pid} = make_pipeline(in_path, out_path)
 
       assert Pipeline.play(pid) == :ok
-      assert_receive :eos, 1000
+      assert_end_of_stream(pid, :sink, :input, 1000)
       assert_files_equal(out_path, ref_path)
     end
   end
