+++ test/encoder/encoder_native_test.exs
@@ -3,7 +3,7 @@ defmodule Encoder.NativeTest do
   alias Membrane.Element.FFmpeg.H264.Encoder.Native, as: Enc
 
   test "Encode 1 240p frame" do
-    in_path = "fixtures/reference-100-240p.raw" |> Path.expand(__DIR__)
+    in_path = "../fixtures/reference-100-240p.raw" |> Path.expand(__DIR__)
 
     assert {:ok, file} = File.read(in_path)
     assert {:ok, ref} = Enc.create(320, 240, :I420, :fast, :high, 30, 1, 23)
