+++ test/transcoding/transcoding_test.exs
@@ -1,9 +1,11 @@
 defmodule TranscodingTest do
-  alias Membrane.Pipeline
+  import Membrane.Testing.Assertions
+  alias Membrane.Testing.Pipeline
+  alias Membrane.Element
   use ExUnit.Case
 
   def prepare_paths(filename) do
-    in_path = "fixtures/input-#{filename}.h264" |> Path.expand(__DIR__)
+    in_path = "../fixtures/input-#{filename}.h264" |> Path.expand(__DIR__)
     out_path = "/tmp/output-transcode-#{filename}.h264"
     File.rm(out_path)
     on_exit(fn -> File.rm(out_path) end)
@@ -11,7 +13,21 @@ defmodule TranscodingTest do
   end
 
   def make_pipeline(in_path, out_path) do
-    Pipeline.start_link(TranscodingPipeline, %{in: in_path, out: out_path, pid: self()}, [])
+    Pipeline.start_link(%Pipeline.Options{
+      elements: [
+        file_src: %Element.File.Source{chunk_size: 40_960, location: in_path},
+        parser: Element.FFmpeg.H264.Parser,
+        decoder: Element.FFmpeg.H264.Decoder,
+        encoder: %Element.FFmpeg.H264.Encoder{preset: :fast, crf: 30},
+        sink: %Element.File.Sink{location: out_path}
+      ],
+      links: %{
+        {:file_src, :output} => {:parser, :input},
+        {:parser, :output} => {:decoder, :input},
+        {:decoder, :output} => {:encoder, :input},
+        {:encoder, :output} => {:sink, :input}
+      }
+    })
   end
 
   describe "TranscodingPipeline should" do
@@ -20,7 +36,7 @@ defmodule TranscodingTest do
 
       assert {:ok, pid} = make_pipeline(in_path, out_path)
       assert Pipeline.play(pid) == :ok
-      assert_receive :eos, 1000
+      assert_end_of_stream(pid, :sink, :input, 1000)
     end
 
     test "transcode 100 240p frames" do
@@ -28,7 +44,7 @@ defmodule TranscodingTest do
 
       assert {:ok, pid} = make_pipeline(in_path, out_path)
       assert Pipeline.play(pid) == :ok
-      assert_receive :eos, 2000
+      assert_end_of_stream(pid, :sink, :input, 2000)
     end
 
     test "transcode 20 360p frames with 422 subsampling" do
@@ -36,7 +52,7 @@ defmodule TranscodingTest do
 
       assert {:ok, pid} = make_pipeline(in_path, out_path)
       assert Pipeline.play(pid) == :ok
-      assert_receive :eos, 2000
+      assert_end_of_stream(pid, :sink, :input, 2000)
     end
   end
 end
